refactor: 전반적인 코드 리팩토링
리팩토링
- repository 패키지 추가
- 사용되지 않는 함수 제거
- nullable 옵션 추가
- 변수명 수정
