+++ src/main/java/nextstep/subway/domain/Line.java
@@ -17,9 +17,10 @@ public class Line extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "down_station_id", nullable = false)
     private Station downStation;
+    @Column(nullable = false)
     private int distance;
 
-    private Line() {}
+    protected Line() {}
 
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
