+++ src/main/java/nextstep/subway/domain/Station.java
@@ -10,8 +10,7 @@ public class Station extends BaseEntity {
     @Column(unique = true)
     private String name;
 
-    public Station() {
-    }
+    protected Station() {}
 
     public Station(String name) {
         this.name = name;
