+++ src/main/java/nextstep/subway/dto/StationResponse.java
@@ -2,26 +2,20 @@ package nextstep.subway.dto;
 
 import nextstep.subway.domain.Station;
 
-import java.time.LocalDateTime;
-
 public class StationResponse {
     private Long id;
     private String name;
-    private LocalDateTime createdDate;
-    private LocalDateTime modifiedDate;
 
     public static StationResponse of(Station station) {
-        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(), station.getModifiedDate());
+        return new StationResponse(station.getId(), station.getName());
     }
 
-    public StationResponse() {
+    private StationResponse() {
     }
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public StationResponse(Long id, String name) {
         this.id = id;
         this.name = name;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
     }
 
     public Long getId() {
@@ -30,13 +24,5 @@ public class StationResponse {
 
     public String getName() {
         return name;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
     }
 }
